+++ docs/table-of-contents.json
@@ -106,9 +106,9 @@
         {
           "title": "Addons",
           "entries": [
-            { "entry": "docs/api-reference/addons/scenegraph/scenegraph-node" },
-            { "entry": "docs/api-reference/addons/scenegraph/model-node" },
-            { "entry": "docs/api-reference/addons/scenegraph/group-node" }
+            { "entry": "docs/api-reference/experimental/scenegraph/scenegraph-node" },
+            { "entry": "docs/api-reference/experimental/scenegraph/model-node" },
+            { "entry": "docs/api-reference/experimental/scenegraph/group-node" }
           ]
         }
       ]
