+++ examples/script/index.html
@@ -4,7 +4,7 @@
       body {margin: 0; padding: 0;}
     </style>
     <script src="https://unpkg.com/@luma.gl/core/dist/dist.min.js"></script>
-    <script src="https://unpkg.com/@luma.gl/addons/dist/dist.min.js"></script>
+    <script src="https://unpkg.com/@luma.gl/experimental/dist/dist.min.js"></script>
     <script src="https://unpkg.com/gl-matrix@3.1.0/gl-matrix-min.js"></script>
   </head>
   <body>
