+++ examples/showcase/gltf/package.json
@@ -12,7 +12,7 @@
     "@loaders.gl/core": "^1.3.0",
     "@loaders.gl/draco": "^1.3.0",
     "@loaders.gl/polyfills": "^1.3.0",
-    "@luma.gl/addons": "^8.0.0-alpha.9",
+    "@luma.gl/experimental": "^8.0.0-alpha.9",
     "@luma.gl/constants": "^8.0.0-alpha.9",
     "@luma.gl/webgl": "^8.0.0-alpha.9",
     "@luma.gl/gltools": "^8.0.0-alpha.9",
