+++ modules/experimental/README.md
@@ -1,3 +1,3 @@
-# @luma.gl/addons
+# @luma.gl/experimental
 
 Code in this module represents early-stage experiments and are not considered officially supported parts of the API. Use at your own risk!
