+++ modules/experimental/src/gltf/create-gltf-objects.js
