+++ modules/experimental/src/gpgpu/histopyramid/histopyramid-shaders.js
