+++ modules/experimental/src/gpgpu/histopyramid/histopyramid.md
